detect/requirements.txt
absl-py==2.0.0
astor==0.8.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
colorama==0.4.6
gast==0.2.2
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.8.0
importlib-metadata==6.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.21.6
opencv-python==4.8.0.76
opt-einsum==3.3.0
pandas==1.3.5
Pillow==9.5.0
protobuf==3.19.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==2.3.0
torch==1.13.1
tqdm==4.66.1
typing_extensions==4.7.1
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.15.0
zipp==3.15.0
